+++ src/main/java/org/manuel/mysportfolio/config/database/MongoConfig.java
@@ -1,6 +1,7 @@
 package org.manuel.mysportfolio.config.database;
 
 import java.time.Year;
+import java.time.YearMonth;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
@@ -29,6 +30,8 @@ public class MongoConfig {
   public MongoCustomConversions customConversions() {
     final List<Converter<?, ?>> converterList = new ArrayList<>();
     converterList.add(new MongoYearFromStringConverter());
+    converterList.add(new MongoYearMonthFromStringConverter());
+    converterList.add(new MongoStringFromYearMonthConverter());
     return new MongoCustomConversions(converterList);
   }
 
@@ -37,6 +40,24 @@ public class MongoConfig {
     @Override
     public Year convert(final String source) {
       return Year.parse(source);
+    }
+  }
+
+  private static final class MongoYearMonthFromStringConverter implements
+      Converter<String, YearMonth> {
+
+    @Override
+    public YearMonth convert(final String source) {
+      return YearMonth.parse(source);
+    }
+  }
+
+  private static final class MongoStringFromYearMonthConverter implements
+      Converter<YearMonth, String> {
+
+    @Override
+    public String convert(final YearMonth source) {
+      return source.toString();
     }
   }