+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Adding CHANGELOG.
+- Adding PUT in Competition controller
 
 ### Changed
 - Fixing bug of query criteria